Softex/codigos/requirements.txt
pandas
SQLAlchemy
pymysql
streamlit
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib